Migrate 'triggerable' test
